Fix sanity check in `unpack_resource`
`from_raw_parts` trusts that the given `len` is correct and thus `items.len()` will always equal
`count`. Moreover, it is UB to give a length that would cause `len * mem::size_of::<T>()` to exceed
the number of bytes for which the data pointer is valid to read.
A sufficiently smart compiler could therefore optimize out even a fixed length check if it was still
placed after the creation of the `items` slice.

Fix all of that by first asserting that the aligned slice is sufficiently long to hold `count`
elements of `T` and only after that actually creating the `items` slice.